requirements.txt
django-cors-headers,
razorpay